requirements.txt
pydantic
pandas
openai
openpyxl
python-dotenv